requirements.txt
Django==4.0.3
gunicorn==20.1.0
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
Pillow==8.2.0
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
whitenoise==6.0.0
asgiref==3.5.0
djangorestframework==3.13.1